refactor: split major components
